+++ src/ruby/models/mouse.rb
@@ -1,0 +1,11 @@
+# The class for dealing with the {Mouse}. Only one mouse at a time is supported.
+class Mouse
+  # @!group Mouse input
+  # Left mouse button.
+  LEFT = 0
+  # Right mouse button.
+  RIGHT = 1
+  # Middle mouse button (scroll wheel).
+  MIDDLE = 2
+  # @!endgroup
+end