+++ test/support/x_do.rb
@@ -3,13 +3,33 @@ module XDo
     linux? && `which xdotool` != ""
   end
 
+  def self.command
+    "xdotool"
+  end
+
   class Mouse
+    BUTTON = {
+      left: 1,
+      middle: 2,
+      right: 3,
+      wheel_up: 4,
+      wheel_down: 5
+    }
+
     def self.move_to(x, y, sync: true)
-      command = "xdotool mousemove "
+      command = "#{XDo.command} mousemove "
       command << "--sync " if sync
       command << "#{x} #{y}"
 
       `#{command}`
+    end
+
+    def self.button_down(button)
+      `#{XDo.command} mousedown #{button}`
+    end
+
+    def self.button_up(button)
+      `#{XDo.command} mouseup #{button}`
     end
   end
 end